Kullanıcı ekleme tamamlandı

+++ Demo.Server/Demo.Server/Demo.Server.Svc/DataContracts/EmployeeDto.cs
@@ -8,7 +8,7 @@ namespace Demo.Server.Svc.DataContracts
 {
 	public class EmployeeDto
 	{
-		public int EmployeeTCKN { get; set; }
+		public long EmployeeTCKN { get; set; }
 		public string EmployeeName { get; set; }
 		public string EmployeeSurname { get; set; }
 		public int Department { get; set; }
